==> Part1.ipynb <==
#Все необходимые импорты
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from collections import Counter
import seaborn as sns
from fitter import Fitter
import warnings
import time
warnings.filterwarnings('ignore')
pd.set_option('display.max_columns', None)

#Вывод информации для одного файла
df_ = pd.read_table('dir/name.txt', delimiter = ',') 
print(df_)
out = pd.read_table('data1.txt', delimiter = ',')
print(out[['', '', '']].iloc[0])

#Вывод описания 
df = pd.read_csv('dataset5.csv')
df.describe(include='all')

#Запись данных в файл
out = pd.read_table('data4.txt', delimiter = ',')
df["Survival"] = out["Survival"]
df["In-hospital_death"] = out["In-hospital_death"]
print(df)
df.to_csv('dataset.csv', index=False)

# Распределение данных по встречаемости
out = pd.read_table('data2.txt', delimiter = ',')
RIDs = out["head"]
n = len(RIDs)
x = Counter()
for rid in RIDs:
    df_ = pd.read_table('set/' + str(rid) + '.txt', delimiter = ',') 
    x = x + Counter(df_["Parameter"])

unique_parameters = list(x.keys())
one_params = []
mean_params = []
rare_params = []
median_params = []
for parameter in unique_parameters:
    if x[parameter] / n > 10:
        median_params.append(parameter)
    elif x[parameter] / n > 1:
        mean_params.append(parameter)
    elif x[parameter] / n < 1:
        rare_params.append(parameter)
    else:
        one_params.append(parameter)
        
feature_list = one_params + rare_params + mean_params + median_params

#Усреднение распределенных данных
df = pd.DataFrame(columns = feature_list, index = range(n))
for i, rid in enumerate(RIDs):
    df_ = pd.read_table('dir3/' + str(rid) + '.txt', delimiter = ',')
    df_edited = pd.DataFrame(0, index = range(1), columns = feature_list)
    for parameter in one_params:
        df_edited[parameter] = df_['Value'].where(df_["Parameter"] == parameter).sum()
    for parameter in rare_params:
        df_edited[parameter] = df_['Value'].where(df_["Parameter"] == parameter).mean()
    for parameter in mean_params:
        df_edited[parameter] = (df_['Value'].where(df_["Parameter"] == parameter)).mean()
    for parameter in median_params: 
        df_edited[parameter] = (df_['Value'].where(df_["Parameter"] == parameter)).median()
    df.loc[i, feature_list] = df_edited.iloc[0].values

# Вывод распределенных параметров и описание датасета
print("rare_params: ", rare_params)
print("mean_params: ", mean_params)
print("one_params6: ", one_params)
print("median_params: ", median_params)
df.columns
df.info()
