==> Part 4.ipynb <==
#Импорт необходимых библиотек
from sklearn.neural_network import MLPClassifier
from sklearn.neighbors import KNeighborsClassifier
from sklearn.svm import SVC
from sklearn.gaussian_process import GaussianProcessClassifier
from sklearn.gaussian_process.kernels import RBF
from sklearn.tree import DecisionTreeClassifier
from sklearn.ensemble import RandomForestClassifier, AdaBoostClassifier
from sklearn.naive_bayes import GaussianNB
from sklearn.discriminant_analysis import QuadraticDiscriminantAnalysis
from sklearn.model_selection import GridSearchCV
from sklearn.metrics import f1_score, accuracy_score, balanced_accuracy_score, precision_score, recall_score
from sklearn import metrics

names = [
    "Nearest Neighbors",
    "Linear SVM",
    "RBF SVM",
    "Gaussian Process",
    "Decision Tree",
    "Random Forest",
    "Neural Net",
    "AdaBoost",
    "Naive Bayes",
    "QDA",
]

classifiers = [
    KNeighborsClassifier(3),
    SVC(kernel="linear", C=0.025),
    SVC(gamma=2, C=1),
    GaussianProcessClassifier(1.0 * RBF(1.0)),
    DecisionTreeClassifier(max_depth=5),
    RandomForestClassifier(max_depth=5, n_estimators=10, max_features=1),
    MLPClassifier(alpha=1, max_iter=1000),
    AdaBoostClassifier(),
    GaussianNB(),
    QuadraticDiscriminantAnalysis(),
]

for name, clf in zip(names, classifiers):
    clf.fit(X_train_cls, y_train_cls)
    pred_cls = clf.predict(X_val_cls)
    score = f1_score(y_val_cls, pred_cls)
    print(name, score)

parameters = {'max_depth':[9, 11, 13], 'n_estimators':[100, 200, 300], 'max_features': ['sqrt', 'log2']}

clf = GridSearchCV(estimator=RandomForestClassifier(), param_grid=parameters, scoring='f1')
clf.fit(X_train_cls, y_train_cls)

results = pd.DataFrame.from_dict(clf.cv_results_)
results = results.sort_values('rank_test_score', ignore_index=True)
print(results['params'][0])
pred_cls = clf.predict(X_val_cls)

print(precision_score(y_val_cls, pred_cls))
print(recall_score(y_val_cls, pred_cls))
print(accuracy_score(y_val_cls, pred_cls))
print(f1_score(y_val_cls, pred_cls))
print(balanced_accuracy_score(y_val_cls, pred_cls))

pred_cls = clf.predict_proba(X_val_cls)
fpr, tpr, _ = metrics.roc_curve(y_val_cls,  pred_cls[:, 1])
auc = metrics.roc_auc_score(y_val_cls, pred_cls[:, 1])

plt.plot(fpr, tpr, label="AUC="+str(auc))
plt.ylabel('True Positive Rate')
plt.xlabel('False Positive Rate')
plt.legend(loc=4)
plt.show()

parameters = {'hidden_layer_sizes':[(20,), (40,), (100,)], 'activation':['identity', 'relu']}

clf = GridSearchCV(estimator=MLPClassifier(), param_grid=parameters, scoring='f1')
clf.fit(X_train_cls, y_train_cls)

results = pd.DataFrame.from_dict(clf.cv_results_)
results = results.sort_values('rank_test_score', ignore_index=True)
print(results['params'][0])
pred_cls = clf.predict(X_val_cls)

print(precision_score(y_val_cls, pred_cls))
print(recall_score(y_val_cls, pred_cls))
print(accuracy_score(y_val_cls, pred_cls))
print(f1_score(y_val_cls, pred_cls))
print(balanced_accuracy_score(y_val_cls, pred_cls))

pred_cls = clf.predict_proba(X_val_cls)
fpr, tpr, _ = metrics.roc_curve(y_val_cls,  pred_cls[:, 1])
auc = metrics.roc_auc_score(y_val_cls, pred_cls[:, 1])

plt.plot(fpr, tpr, label="AUC="+str(auc))
plt.ylabel('True Positive Rate')
plt.xlabel('False Positive Rate')
plt.legend(loc=4)
plt.show()

parameters = {'var_smoothing':[1e-9, 1e-7, 1e-6, 1e-5]}

clf = GridSearchCV(estimator=GaussianNB(), param_grid=parameters, scoring='f1')
clf.fit(X_train_cls, y_train_cls)

results = pd.DataFrame.from_dict(clf.cv_results_)
results = results.sort_values('rank_test_score', ignore_index=True)
print(results['params'][0])
pred_cls = clf.predict(X_val_cls)

print(precision_score(y_val_cls, pred_cls))
print(recall_score(y_val_cls, pred_cls))
print(accuracy_score(y_val_cls, pred_cls))
print(f1_score(y_val_cls, pred_cls))
print(balanced_accuracy_score(y_val_cls, pred_cls))

pred_cls = clf.predict_proba(X_val_cls)
fpr, tpr, _ = metrics.roc_curve(y_val_cls,  pred_cls[:, 1])
auc = metrics.roc_auc_score(y_val_cls, pred_cls[:, 1])

plt.plot(fpr, tpr, label="AUC="+str(auc))
plt.ylabel('True Positive Rate')
plt.xlabel('False Positive Rate')
plt.legend(loc=4)
plt.show()

# Прогноз значения целеовй переменной, качество модели
clf = RandomForestClassifier(max_depth=9, n_estimators=300, max_features='log2')
clf.fit(X_train_cls, y_train_cls)
pred_cls = clf.predict(X_test_cls)

print(precision_score(y_test_cls, pred_cls))
print(recall_score(y_test_cls, pred_cls))
print(accuracy_score(y_test_cls, pred_cls))
print(f1_score(y_test_cls, pred_cls))
print(balanced_accuracy_score(y_test_cls, pred_cls))

pred_cls = clf.predict_proba(X_val_cls)
fpr, tpr, _ = metrics.roc_curve(y_val_cls,  pred_cls[:, 1])
auc = metrics.roc_auc_score(y_val_cls, pred_cls[:, 1])

plt.plot(fpr, tpr, label="AUC="+str(auc))
plt.ylabel('True Positive Rate')
plt.xlabel('False Positive Rate')
plt.legend(loc=4)
plt.show()
